Casting some values to float in super() call in constructor.
git-svn-id: http://chianti.ucsd.edu/svn@3585 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/fung/src/cytoscape/fung/EdgeLabel.java
@@ -27,7 +27,7 @@ public final class EdgeLabel extends Label
                    final byte justify)
   {
     super(text, font, scaleFactor, paint, textAnchor,
-          offsetVector.getX(), offsetVector.getY(), justify);
+          (float) offsetVector.getX(), (float) offsetVector.getY(), justify);
     m_edgeAnchor = edgeAnchor;
     switch (m_edgeAnchor) {
     case EDGE_ANCHOR_MIDPOINT:
